+++ packages/react-router/__tests__/Route-test.tsx
@@ -22,6 +22,25 @@ describe("A <Route>", () => {
     `);
   });
 
+  it("renders its `Component` prop", () => {
+    let renderer: TestRenderer.ReactTestRenderer;
+    TestRenderer.act(() => {
+      renderer = TestRenderer.create(
+        <MemoryRouter initialEntries={["/home"]}>
+          <Routes>
+            <Route path="home" Component={() => <h1>Home</h1>} />
+          </Routes>
+        </MemoryRouter>
+      );
+    });
+
+    expect(renderer.toJSON()).toMatchInlineSnapshot(`
+      <h1>
+        Home
+      </h1>
+    `);
+  });
+
   it("renders its child routes when no `element` prop is given", () => {
     let renderer: TestRenderer.ReactTestRenderer;
     TestRenderer.act(() => {