+++ packages/router/jest.config.js
@@ -1,3 +1,4 @@
+/** @type {import('jest').Config} */
 module.exports = {
   testEnvironment: "jsdom",
   testMatch: ["**/__tests__/*-test.(js|ts)"],
@@ -15,4 +16,5 @@ module.exports = {
       "@web3-storage/multipart-parser"
     ),
   },
+  reporters: ["default"],
 };