+++ packages/router/CHANGELOG.md
@@ -1,5 +1,12 @@
 # `@remix-run/router`
 
+## 1.6.1
+
+### Patch Changes
+
+- Fix `basename` handling when navigating without a path ([#10433](https://github.com/remix-run/react-router/pull/10433))
+- "Same hash" navigations no longer re-run loaders to match browser behavior (i.e. `/path#hash -> /path#hash`) ([#10408](https://github.com/remix-run/react-router/pull/10408))
+
 ## 1.6.0
 
 ### Minor Changes